Moved the confirmation dialog class to where it is used.
