+++ frontend/src/modules/report/FeatureListContainer/ConfirmationDialog/ConfirmationDialog.tsx
